% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_metadata.R
\name{browse_metadata}
\alias{browse_metadata}
\title{browse_metadata}
\usage{
browse_metadata(json_file = NULL, output_dir = NULL)
}
\arguments{
\item{json_file}{The metadata file. This should be a json download from the
metadata catalogue. By default, the .json file within 'inst/inputs' is used.}

\item{output_dir}{The path to the directory where the three output files
will be saved. By default, the current working directory is used.}
}
\value{
The function will return three files, 'BROWSE_table...html',
'BROWSE_bar...html' and 'BROWSE_bar...csv' which gives summary information
for this dataset. Open the two html output files in your browser and use
these as reference when running the map_metadata function.
}
\description{
Run this function before map_metadata \cr \cr
This function will read in the metadata file for a chosen dataset and save
three summary outputs. The first is a table output, storing the name and
description of the dataset, and each table within it. The second is a bar
chart, summarising how many variables there are for each table, and whether
these variables have a missing description. The third is a csv file storing
the data that created this bar chart. \cr \cr
}
\examples{
# Demo run requires no function inputs or user interaction
browse_metadata()
}
