% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseMetadata-package.R
\docType{package}
\name{browseMetadata-package}
\alias{browseMetadata}
\alias{browseMetadata-package}
\title{browseMetadata: Browse and categorise health metadata}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This R package allows researchers to explore publicly available metadata from health datasets. It is a tool to help researchers plan projects by interacting with metadata prior to gaining full access to health datasets. At the early stages of a project, researchers can use this tool to browse datasets and categorise variables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aim-rsf.github.io/browseMetadata/}
}

}
\author{
\strong{Maintainer}: Rachael Stickland \email{rstickland@turing.ac.uk} (\href{https://orcid.org/0000-0003-3398-4272}{ORCID})

}
\keyword{internal}
