% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseMetadata.R
\name{browseMetadata}
\alias{browseMetadata}
\title{browseMetadata}
\usage{
browseMetadata(json_file, output_dir = NULL)
}
\arguments{
\item{json_file}{The metadata file. This should be downloaded from the
metadata catalogue as a json file. Use
'data-raw/maternity_indicators_dataset_(mids)_20240105T132210.json' for an
example download.}

\item{output_dir}{The path to the directory where the two output files
will be saved. By default, the current working directory is used.}
}
\value{
The function will return two files, 'BROWSE_table_' and 'BROWSE_bar'
which gives summary informatin for this dataset and can be used as reference
when running the MapMetadata function. Open these outputs in a browser.
}
\description{
Run this function before MapMetadata. \cr \cr
This function will read in the metadata file for a chosen dataset and save
two summary outputs. The first is a table output, storing the name and
description of the dataset, and each table within it. The second is a bar
chart, summarising how many variables there are for each table, and whether
these variables have a missing description. \cr \cr
}
