% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseMetadata_prompt_from_list.R
\name{browseMetadata_prompt_from_list}
\alias{browseMetadata_prompt_from_list}
\title{DOCUMENT!}
\usage{
browseMetadata_prompt_from_list(
  pre_prompt_df,
  pre_prompt_df_rows,
  prompt_text,
  list_allowed
)
}
\description{
DOCUMENT!
}
