% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseMetadata_user_prompt.R
\name{browseMetadata_user_prompt}
\alias{browseMetadata_user_prompt}
\title{DOCUMENT!}
\usage{
browseMetadata_user_prompt(pre_prompt_text = NULL, prompt_text, any_keys)
}
\description{
DOCUMENT!
}
