% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_prompt_list.R
\name{browseMetadata_user_prompt_list}
\alias{browseMetadata_user_prompt_list}
\title{DOCUMENT!}
\usage{
browseMetadata_user_prompt_list(
  pre_prompt_df,
  pre_prompt_df_rows,
  prompt_text,
  list_allowed
)
}
\description{
DOCUMENT!
}
