% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{concensus_on_mismatch}
\alias{concensus_on_mismatch}
\title{concensus_on_mismatch}
\usage{
concensus_on_mismatch(ses_join, Table_df, datavar, domain_code_max)
}
\arguments{
\item{ses_join}{The joined dataframes from the two sessions}

\item{Table_df}{Metadata from the json file, for one table in the dataset}

\item{datavar}{Data Element n}

\item{domain_code_max}{The maximum allowable domain code integer}
}
\value{
It returns a list of 2: the domain code and the note from the consensus decision
}
\description{
This function is called within the map_metadata_compare function. \cr \cr
For a specific data element, it compares the domain code categorisation between two sessions.
If the categorisation differs, it prompts the user for a new consensus decision by presenting the json metadata. \cr \cr
}
