% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_output.R
\name{convert_output}
\alias{convert_output}
\title{convert_output}
\usage{
convert_output(output_csv, output_dir)
}
\arguments{
\item{output_csv}{The name of the 'OUTPUT_' csv file that was created from domain_mapping.R}

\item{output_dir}{The location of output_csv}
}
\value{
The function will return 'L-OUTPUT_' in the same output_dir
}
\description{
The 'OUTPUT_' file groups multiple categorisations onto one line e.g. '1,3' \cr \cr
This function creates a new longer output 'L-OUTPUT_' which gives each categorisation its own row \cr \cr
This 'L-OUTPUT_' may be more useful when using these csv files in later analyses
}
