% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_inputs.R
\name{copy_previous}
\alias{copy_previous}
\title{copy_previous}
\usage{
copy_previous(dataset_name, output_dir)
}
\arguments{
\item{dataset_name}{Name of the dataset}

\item{output_dir}{Output directory to be searched}
}
\value{
It returns a list of 2: df_prev_exist (a boolean) and df_prev (NULL or populated with data elements to copy)
}
\description{
Internal Function: This function is called within the map_metadata function. \cr \cr
It searches for previous OUTPUT files in the output_dir, that match the dataset name. \cr \cr
If files exist, it removes duplicates and autos, and stores the rest of the data elements in a dataframe. \cr \cr
}
\keyword{internal}
