% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_empty_desc.R
\name{count_empty_desc}
\alias{count_empty_desc}
\title{count_empty_desc}
\usage{
count_empty_desc(Table_df, Table_colname)
}
\arguments{
\item{Table_df}{Output from json_table_to_df.R}
}
\value{
A simpler dataframe with Y/N empty counts for variables in the table.
}
\description{
This function is called within the browseMetadata function. \cr \cr
It reads in a data frame that summarises one table of the dataset. \cr \cr
It counts missing variable descriptions, based on specified criteria.
}
