% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-package_data.R
\docType{data}
\name{domain_list}
\alias{domain_list}
\title{Internal: List of Domains}
\format{
A data frame with 5 rows and 1 column
}
\source{
The csv was manually created
}
\usage{
data(domain_list)
}
\description{
Internal Dataset: A simplified list of domains, to demo the function map_metadata.R \cr \cr
This data was created with these two steps:
\enumerate{
 \item \code{domain_list <- read.csv('browseMetadata/inst/inputs/domain_list_demo.csv',header=FALSE)}
 \item \code{usethis::use_data(domain_list)}
}
}
\keyword{internal}
