% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{end_plot}
\alias{end_plot}
\title{Internal: end_plot}
\usage{
end_plot(df, table_name, ref_table)
}
\arguments{
\item{df}{The Output dataframe with all the domain categorisations}

\item{table_name}{The table name}

\item{ref_table}{The domain code reference table (which domain maps to which integer)}
}
\value{
It returns a ggplot
}
\description{
This function is called within the map_metadata function. \cr \cr
A summary plot is created that includes the domain code reference table and counts of domain code categorisations \cr \cr
}
\keyword{internal}
