% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{json_table_to_df}
\alias{json_table_to_df}
\title{json_table_to_df}
\usage{
json_table_to_df(dataset, n)
}
\arguments{
\item{dataset}{This is the dataModel field of the json}

\item{n}{The table number (as a json can have multiple tables within a dataset)}
}
\value{
A dataframe for that specific table, including data label, description and type.
}
\description{
Internal Function: This function is called within the browse_metadata and map_metadata functions. \cr \cr
It reads in the nested lists from the json and extracts information needed into a dataframe. \cr \cr
It does this for one specific table in a dataset. \cr \cr
}
\keyword{internal}
