% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-package_data.R
\docType{data}
\name{log_output}
\alias{log_output}
\title{Output log dataframe}
\format{
A data frame with 1 empty row and 9 columns
}
\source{
The dataframe was manually created as package data, using the above code.
}
\usage{
data(log_output)
}
\description{
Empty log output dataframe for map_metadata.R to fill. Created by: \cr \cr
\enumerate{
 \item \code{log_output <- data.frame(timestamp = character(1),
 browseMetadata = character(1),
 Initials = character(1),
 MetaDataVersion = character(1),
 MetaDataLastUpdated = character(1),
 DomainListDesc = character(1),
 Dataset = character(1),
 Table = character(1),
 Table_note = character(1))}
 \item \code{usethis::use_data(log_output)}
}
}
\keyword{datasets}
