% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-package_data.R
\docType{data}
\name{log_output_df}
\alias{log_output_df}
\title{Internal: Output log dataframe}
\format{
A data frame with 1 empty row and 9 columns
}
\source{
The dataframe was manually created as package data, using the above code.
}
\usage{
data(log_output_df)
}
\description{
Internal Dataset: Empty log output dataframe for map_metadata.R to fill. Created by: \cr \cr
\enumerate{
 \item \code{log_output_df <- data.frame(timestamp = character(1),
 browseMetadata = character(1),
 initials = character(1),
 metadata_version = character(1),
 metadata_last_updated = character(1),
 domain_list_desc = character(1),
 dataset = character(1),
 table = character(1),
 table_note = character(1))}
 \item \code{usethis::use_data(log_output_df)}
}
}
\keyword{internal}
