% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-package_data.R
\docType{data}
\name{look_up}
\alias{look_up}
\title{Auto-categorisations}
\format{
A data frame with a variable number of rows and 3 columns
}
\source{
The csv was manually created
}
\usage{
data(look_up)
}
\description{
A list of pre-defined pairings between data element and domain code. \cr \cr
For each data element that map_metadata processes: \cr \cr
If it is contained within this look-up table, it uses the auto-categorised domain code rather than asking the user to categorise.\cr\cr
This data was created with these two steps:
\enumerate{
 \item \code{look_up <- read.csv('browseMetadata/inst/inputs/look_up.csv')}
 \item \code{usethis::use_data(look_up)}
}
}
\keyword{datasets}
