% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMetadata_convert_outputs.R
\name{mapMetadata_convert_outputs}
\alias{mapMetadata_convert_outputs}
\title{mapMetadata_convert_outputs}
\usage{
mapMetadata_convert_outputs(output_csv, output_dir)
}
\arguments{
\item{output_csv}{The name of the 'OUTPUT_' csv file that was created from mapMetadata.R}

\item{output_dir}{The location of output_csv}
}
\value{
The function will return 'L-OUTPUT_' in the same output_dir
}
\description{
The 'OUTPUT_' file groups multiple categorisations onto one line e.g. '1,3' \cr \cr
This function creates a new longer output 'L-OUTPUT_' which gives each categorisation its own row \cr \cr
This 'L-OUTPUT_' may be more useful when using these csv files in later analyses
}
