% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-package_data.R
\docType{data}
\name{output_df}
\alias{output_df}
\title{Internal: Output dataframe}
\format{
A data frame with 0 rows and 6 columns
}
\source{
The dataframe was manually created as package data, using the above code.
}
\usage{
data(output_df)
}
\description{
Internal Dataset: Empty output dataframe for map_metadata.R to fill. Created by: \cr \cr
\enumerate{
 \item \code{output_df <- data.frame(timestamp = character(0),
 table = character(0),
 data_element_n = character(0),
 data_element = character(0),
 domain_code = character(0),
 note = character(0))}
 \item \code{usethis::use_data(output_df)}
}
}
\keyword{internal}
