% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_categorisation.R
\name{user_categorisation}
\alias{user_categorisation}
\title{user_categorisation}
\usage{
user_categorisation(data_element, data_desc, data_type, domain_code_max)
}
\arguments{
\item{data_element}{Name of the variable}

\item{data_desc}{Description of the variable}

\item{data_type}{Data type of the variable}

\item{domain_code_max}{Max code in the domain list (0-3 auto included, then N included via domain_file)}
}
\value{
It returns a list containing the decision and decision note
}
\description{
This function is used within the domain_mapping function. \cr \cr
It displays data properties to the user and requests a categorisation into a domain. \cr \cr
An optional note can be included with the categorisation.
}
