% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interactions.R
\name{user_categorisation_loop}
\alias{user_categorisation_loop}
\title{user_categorisation_loop}
\usage{
user_categorisation_loop(
  start_v,
  end_v,
  table_df,
  df_prev_exist,
  df_prev,
  lookup,
  df_plots,
  output
)
}
\arguments{
\item{start_v}{Index of data element to start}

\item{end_v}{Index of data element to end}

\item{table_df}{Dataframe with the table information, extracted from json metadata}

\item{df_prev_exist}{Boolean to indicate with previous dataframes exist (to copy from)}

\item{df_prev}{Previous dataframes to copy from (or NULL)}

\item{lookup}{The lookup table to enable auto categorisations}

\item{df_plots}{Output from the ref_plot function, to indicate maximum domain code allowed}

\item{output}{Empty output dataframe, to fill}
}
\value{
An output dataframe containing information about the table, data elements and categorisations
}
\description{
This function is called within the map_metadata function. \cr \cr
Given a specific table and a number of data elements to search, it checks for 3 different sources of domain categorisation: \cr \cr
1 - If data elements match those in the look-up table, auto categorise them \cr \cr
2 - If data elements match to previous table output, copy them \cr \cr
3 - If no match for 1 or 2, data elements are categorised by the user \cr \cr
}
