% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_prompt.R
\name{user_prompt}
\alias{user_prompt}
\title{user_prompt}
\usage{
user_prompt(prompt_text, any_keys)
}
\arguments{
\item{prompt_text}{Text to display to the user, to prompt their response.}

\item{any_keys}{Boolean to determine if any key responses are allowable.
If FALSE, only these are allowed: Y, y, N and n.}
}
\value{
It returns variable text, depending on any_keys.
}
\description{
This function is called within the mapMetadata function. \cr \cr
It prompts a response from the user. \cr \cr
}
