% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interactions.R
\name{user_prompt_list}
\alias{user_prompt_list}
\title{user_prompt_list}
\usage{
user_prompt_list(prompt_text, list_allowed, empty_allowed)
}
\arguments{
\item{prompt_text}{Text to display to the user, to prompt their response.}

\item{list_allowed}{A list of allowable integer responses.}

\item{empty_allowed}{A boolean specifying if no response is allowed.}
}
\value{
It returns a list of integers to process, that match the prompt options.
}
\description{
This function is called within the map_metadata function. \cr \cr
It prompts a response from the user, in the form of a list. \cr \cr
It checks if the user has given the an input within the allowed range - if not, it re-sends prompt. \cr \cr
}
