% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_comparison.R
\name{valid_comparison}
\alias{valid_comparison}
\title{valid_comparison
This function is called within the browseMetadata_compare_outputs function. \cr \cr
It reads two inputs to see if they are equal. \cr \cr
If the test is 'warning' status and inputs are not equal it gives warning but continues. \cr \cr
If the test is 'danger' status and inputs are not equal it stops and exits, with error message. \cr \cr}
\usage{
valid_comparison(input1, input2, severity, severity_text)
}
\arguments{
\item{input1}{Input 1}

\item{input2}{Input 2}

\item{severity}{Level of severity. Only 'danger' or 'warning'}

\item{severity_text}{The text to print if inputs are not equal.}
}
\value{
It returns variable text, depending on any_keys.
}
\description{
valid_comparison
This function is called within the browseMetadata_compare_outputs function. \cr \cr
It reads two inputs to see if they are equal. \cr \cr
If the test is 'warning' status and inputs are not equal it gives warning but continues. \cr \cr
If the test is 'danger' status and inputs are not equal it stops and exits, with error message. \cr \cr
}
