% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseMetadata-package.R
\docType{package}
\name{browseMetadata-package}
\alias{browseMetadata}
\alias{browseMetadata-package}
\title{browseMetadata}
\description{
Browses available metadata, to catergorise/label each variable in a dataset.
}
\author{
\strong{Maintainer}: Rachael Stickland \email{rstickland@turing.ac.uk} (\href{https://orcid.org/0000-0003-3398-4272}{ORCID})

}
\keyword{internal}
