% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseMetadata-package.R
\docType{package}
\name{browseMetadata-package}
\alias{browseMetadata}
\alias{browseMetadata-package}
\title{browseMetadata: Browse and categorise health metadata}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Visualise and categorise publicly available metadata from health datasets. By interacting with metadata prior to gaining full access to health datasets, researchers can use this tool to browse datasets and categorise variables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aim-rsf.github.io/browseMetadata/}
  \item Report bugs at \url{https://github.com/aim-rsf/browseMetadata/issues}
}

}
\author{
\strong{Maintainer}: Rachael Stickland \email{rstickland@turing.ac.uk} (\href{https://orcid.org/0000-0003-3398-4272}{ORCID})

Other contributors:
\itemize{
  \item Batool Almarzouq (\href{https://orcid.org/0000-0002-3905-2751}{ORCID}) [contributor]
  \item Mahwish Mohammad (\href{https://orcid.org/0009-0004-5295-0726}{ORCID}) [contributor]
  \item Daniel Delbarre (\href{https://orcid.org/0000-0003-4633-4252}{ORCID}) [contributor]
  \item Nida Ziauddeen (\href{https://orcid.org/0000-0002-8964-5029}{ORCID}) [contributor]
}

}
\keyword{internal}
