% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_csv_outputs.R
\name{compare_csv_outputs}
\alias{compare_csv_outputs}
\title{compare_csv_outputs}
\usage{
compare_csv_outputs(csv_file_1, csv_file_2, json_file)
}
\arguments{
\item{csv_file_1}{CSV output file from running domain_mapping}

\item{csv_file_2}{CSV output file from running domain_mapping (different to csv_file_1)}

\item{json_file}{The metadata file used when running domain_mapping (should be the same for csv_file_1 and csv_file_2)}
}
\value{
It returns csv_3, with consensus decisions
}
\description{
This function is to be used after running the domain_mapping function. \cr \cr
It compares two csv outputs, finds their differences, and asks for a consensus. \cr \cr
}
