% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_csv_outputs.R
\name{compare_csv_outputs}
\alias{compare_csv_outputs}
\title{compare_csv_outputs}
\usage{
compare_csv_outputs(
  session_dir,
  session1_base,
  session2_base,
  json_file,
  domain_file
)
}
\arguments{
\item{session_dir}{This directory should contain 2 csv files for each session (LOG_ and OUTPUT_), 4 csv files in total.}

\item{session1_base}{Base file name for session 1 e.g. 'NationalCommunityChildHealthDatabase(NCCHD)_BLOOD_TEST_2024-07-05_16-07-38.599493'}

\item{session2_base}{Base file name for session 1 e.g. 'NationalCommunityChildHealthDatabase(NCCHD)_BLOOD_TEST_2024-07-08_12-03-30.429336'}

\item{json_file}{The full path to the metadata file used when running domain_mapping (should be the same for session 1 and session 2)}

\item{domain_file}{The full path to the domain file used when running domain_mapping (should be the same for session 1 and session 2)}
}
\value{
It returns a csv output, which represents the consensus decisions between session 1 and session 2
}
\description{
This function is to be used after running the domain_mapping function. \cr \cr
It compares outputs from two sessions, finds their differences, and asks for a consensus. \cr \cr
}
