% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_previous.R
\name{copy_previous}
\alias{copy_previous}
\title{copy_previous}
\usage{
copy_previous(Dataset_Name, output_dir)
}
\arguments{
\item{Dataset_Name}{Name of the dataset}

\item{output_dir}{Output directory to be searched}
}
\value{
It returns a list of 2: df_prev_exist (a boolean) and df_prev (NULL or populated with data elements to copy)
}
\description{
This function is called within the mapMetadata function. \cr \cr
It searches for previous OUTPUT files in the output_dir, that match the dataset name. \cr \cr
If files exist, it removes duplicates and autos, and stores the rest of the data elements in a dataframe. \cr \cr
}
