% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_mapping.R
\name{domain_mapping}
\alias{domain_mapping}
\title{domain_mapping}
\usage{
domain_mapping(json_file, domain_file, demo_mode = FALSE)
}
\arguments{
\item{json_file}{The metadata file. This should be downloaded from the metadata catalogue as a json file.}

\item{domain_file}{The file that lists the domains of interest to be used within the research study, provided as a csv with each domain on a separate line, within quotations.}

\item{demo_mode}{Write TRUE to run the function in demo mode and leave the other inputs blank. This uses the demo data in the package - see example below.}
}
\value{
The function will return a log file with your mapping between variables and domains, alongside details about the Data Asset.
}
\description{
This function will read in the meta-data of a Data Asset obtained from metadata catalogue: https://modelcatalogue.cs.ox.ac.uk/hdruk_live.
It will loop through all the Data Elements (variable names) in each Data Class (table), and ask you to categorise each variable into one of your chosen domains.
Information about the Data Asset and Data Class can be displayed to the command window for reference.
The domains will appear in the Plots tab, with the labels you should use for the categorisation.
}
\examples{
# Run the code with the demo files
\emph{domain_mapping(,,TRUE)}
# Respond with your initials when prompted.
# Respond 'Demo List.' for the description of demo list.
# Respond 'Y' if you want to see the descriptions printed out.
# Respond '1,10' to the RANGE OF VARIABLES prompt (or process the full 93 variables if you like!)
# Reference the plot tab and categorise each variable into a single ('1') or multiple  ('1,2') domain.
# Write a note explaining your category choice (optional).
}
