% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_table_to_df.R
\name{json_table_to_df}
\alias{json_table_to_df}
\title{join_outputs}
\usage{
json_table_to_df(Dataset, n)
}
\arguments{
\item{Dataset}{This is the dataModel field of the json}

\item{n}{The Dataset number (as a json can have multiple datasets)}
}
\value{
A dataframe for that specific table, including data label, description and type.
}
\description{
This function is called within the browse_metadata and map_metadata functions. \cr \cr
It reads in the nested lists from the json and extracts information needed into a dataframe. \cr \cr
It does this for one specific table in a dataset. \cr \cr
}
