% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_metadata_compare.R
\name{map_metadata_compare}
\alias{map_metadata_compare}
\title{map_metadata_compare}
\usage{
map_metadata_compare(
  session_dir,
  session1_base,
  session2_base,
  json_file,
  domain_file
)
}
\arguments{
\item{session_dir}{This directory should contain 2 csv files for each session (LOG_ and OUTPUT_), 4 csv files in total.}

\item{session1_base}{Base file name for session 1 e.g. 'NationalCommunityChildHealthDatabase(NCCHD)_BLOOD_TEST_2024-07-05-16-07-38'}

\item{session2_base}{Base file name for session 1 e.g. 'NationalCommunityChildHealthDatabase(NCCHD)_BLOOD_TEST_2024-07-08-12-03-30'}

\item{json_file}{The full path to the metadata file used when running map_metadata (should be the same for session 1 and session 2)}

\item{domain_file}{The full path to the domain file used when running map_metadata (should be the same for session 1 and session 2)}
}
\value{
It returns a csv output, which represents the consensus decisions between session 1 and session 2
}
\description{
This function is to be used after running the map_metadata function. \cr \cr
It compares csv outputs from two sessions, finds their differences, and asks for a consensus. \cr \cr
}
