% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_metadata_convert.R
\name{map_metadata_convert}
\alias{map_metadata_convert}
\title{map_metadata_convert}
\usage{
map_metadata_convert(output_csv, output_dir)
}
\arguments{
\item{output_csv}{The name of the 'OUTPUT_' csv file that was created from map_metadata}

\item{output_dir}{The location of output_csv}
}
\value{
The function will return 'L-OUTPUT_' in the same output_dir
}
\description{
The 'OUTPUT_' file groups multiple categorisations onto one line e.g. Domain_code could read '1,3' \cr \cr
This function creates a new longer output 'L-OUTPUT_' which gives each categorisation its own row. \cr \cr
This 'L-OUTPUT_' may be more useful when using these csv files in later analyses.
}
\examples{
# Create a temporary directory for the example
temp_dir <- tempdir()

# Create an example OUTPUT_ CSV file with primary care data
example_data <- data.frame(
  patient_id = c(101, 102, 103, 104),
  condition_code = c("A01", "B02", "A01,C03", "D04"),
  medication = c("Medication1", "Medication2", "Medication3", "Medication4")
)
write.csv(example_data, file = file.path(temp_dir, "OUTPUT_example.csv"), row.names = FALSE)

# Use the function to process the example file
map_metadata_convert("OUTPUT_example.csv", temp_dir)

# Check the output
read.csv(file.path(temp_dir, "L-OUTPUT_example.csv"))
}
