% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_metadata_convert.R
\name{map_metadata_convert}
\alias{map_metadata_convert}
\title{map_metadata_convert}
\usage{
map_metadata_convert(output_csv, output_dir)
}
\arguments{
\item{output_csv}{The name of the 'OUTPUT_' csv file that was created from map_metadata}

\item{output_dir}{The location of output_csv}
}
\value{
The function will return 'L-OUTPUT_' in the same output_dir
}
\description{
The 'OUTPUT_' file groups multiple categorisations onto one line e.g. Domain_code could read '1,3' \cr \cr
This function creates a new longer output 'L-OUTPUT_' which gives each categorisation its own row. \cr \cr
This 'L-OUTPUT_' may be more useful when using these csv files in later analyses.
}
\examples{
# Locate file path and file name for the example files in the package
demo_output_dir <- system.file("outputs", package = "browseMetadata")
demo_output_csv <- "OUTPUT_NationalCommunityChildHealthDatabase(NCCHD)_CHILD_2024-11-27-14-19-55.csv"

# Run the function
map_metadata_convert(output_csv = demo_output_csv, output_dir = demo_output_dir)
}
