% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Output.R
\docType{data}
\name{Output}
\alias{Output}
\title{Output dataframe}
\format{
A data frame with 0 rows and 6 columns
}
\source{
The dataframe was manually created for the mapMetadata.R function
}
\usage{
data(Output)
}
\description{
Empty output dataframe for mapMetadata.R to fill. \cr \cr
}
\details{
\item \code{Output <- data.frame(
timestamp = character(0),
Table = character(0),
DataElement_N = character(0),
DataElement = character(0),
Domain_code = character(0),
Note = character(0)
)}

\item \code{usethis::use_data(Output)}
}
\keyword{datasets}
