% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_comparison.R
\name{valid_comparison}
\alias{valid_comparison}
\title{valid_comparison}
\usage{
valid_comparison(input1, input2, severity, severity_text)
}
\arguments{
\item{input1}{Input 1}

\item{input2}{Input 2}

\item{severity}{Level of severity. Only 'danger' or 'warning'}

\item{severity_text}{The text to print if inputs are not equal.}
}
\value{
Returns nothing if inputs are equal. If inputs are not equal, returns variable text depending on level of severity.
}
\description{
This function is called within the map_metadata_compare function. \cr \cr
It reads two inputs to see if they are equal. \cr \cr
If the test is 'warning' status and inputs are not equal it gives warning but continues. \cr \cr
If the test is 'danger' status and inputs are not equal it stops and exits, with error message. \cr \cr
}
